requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.0
django-braces==1.15.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-import-export==2.7.1
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-model-utils==4.3.1
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-utils-six==2.0
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
idna==3.4
jdatetime==4.1.1
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.7.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
swapper==1.3.0
tablib==3.4.0
tzdata==2023.3
urllib3==2.0.2
xlrd==2.0.1
xlwt==1.3.0
